Add a reset-on-restart handler.
This fixes a bug that I added. We had a test app that had
to be killed after each run; I got so used to that, that
I did it for every GVRF app, and so didn't notice when my
singletons were causing problems.

This patch adds a reset-on-restart mechanism, and uses
it to reset all statics as necessary.

All samples now restart just fine whether you exit via
the Home key or the Back key; you can restart an app
from the Task Switcher.