Fix SceneObject::removeChildObject() parent reset
The current code, by calling lock() on _parent (which is a weak_ptr),
creates a temporary shared_ptr instance, which it then calls reset()
on.  While this does reset() that temporary shared_ptr, it doesn't do
anything to _parent itself, which continues to reference the old
parent until a new parent is set.

This change calls reset on _parent itself, which clears the reference
to the parent scene object.