Clean up post-effect handling.
The Renderer::renderCamera() method 'asks' the camera for
the current list of post effects; it ALSO processes a
list of 'extra' post effect data, passed down from the
Java layer. This is wasteful, as the Java layer just
passes an empty List; this also sems buggy, as
Renderer::renderCamera appends any data on the 'extra'
list to each eye.

This appears to be a bit of archaeology. Clearing it out
reduces memory churn, and eliminates two GetLongArray /
ReleaseLongArray pairs, per frame.