Introduce class Preconditions and restructure some parameter checks in GVRMesh, GVRMaterial, and GVRCameraRig
One of our goal is to provide Java side exception as possible as we can while in the illegal states.
Previously I impelment a GVRJniException class which can simulate Java stack trace and throw a JNI crash finally.
I am thinking if we should avoid using this class as possible as we can. So for the readability and further robustness
in GVRF, I want to introduce the class Preconditions and provide some convenient pre-condition check at Java level
so that we can avoid using GVRJniException when crash happens in JNI side.

But I think the use of Preconditions may need to be very careful since unnecessary checks will eventually turn out to be
 a significant overhead. For example, in GVRMaterial and GVRCameraRig I finally decided to only add NaN check in setFloat
is because only in the case of single float parameter may need NaN as a special flag to inform Java that an illegal state
happens. Corresponsively, in the case of float vector/array, we can simply use return pointer back from JNI to Java as an
illegal state identifier. I mean, we probably can't check all illegal states using Preconditions. But in necessary cases,
I think we'd better make use of it and in the future extend it in a proper way.