Fix flickering and allow user to set/get refresh frequency of this GVRTextViewSceneObject.
Fix flickering while more than a certain number of GVRTextViewSceneObject within a Scene.
Actually the problem is caused by the limitation of SurfaceTexture's performance.
If we are trying to refresh many SurfaceTextures within a frame, then it will cause latency issue
which finally turns out to be the flickering.
The patch is something like a load balancer so that we distribute the refresh() calls into interval
frames instead of calling them within one frame. And only if the textview is changed, we trigger the
refresh().

The default interval is 30 and I changed it to 20, and allow user to set the frequency of 3 levels based
on their needs.