Add bounding sphere info to the mesh.
Adds bounding sphere info for the mesh.  This bounding sphere is based
on the bounding box that we're already using for frustum culling.  As
such, the sphere will likely be a little larger than necessary, but will
be accurate enough for now.

The bounding sphere (at least the center of it) will be used in upcoming
check-ins dealing with transparency sorting, and LOD.