Added support for an installable external renderer
This allows an external renderer outside gvrf to act
as a shader and participate in the rendering of the scene.
The external renderer receives the bounding box of the mesh
to be rendered, the projection matrix, and data received
from the application.
The main texture of the material must be a new
ExternalRendererTexture when using an external renderer.