Modified gvr-vuforia sample to demo camera pass-through with vuforia.
The previous sample is modified to remove the earlier used mesh and
textures that show a cube environment. Instead a teapot mesh and
texture are added.

Vuforia related files are heavily modified to initialize and shutdown
Vuforia in a cleaner way.

Also a camera pass through backgound is added and a marker can be tracked
using the camera. The teapot is displayed on top of the marker.