Minimal 360 Photo Viewer
Minimal 360 Photo Viewer that quickly got to be not so minimal.  Tried
to add handling of intent sharing.  This sample seems to have severe
lifecycle issues.  You can run it once, an then never again.  All you
get is black on subsequent runs.  It seems the GVRScript's onInit()
never gets called in the subsequent runs.