Refactor Throttler for dependency injection, type checking and readability
Type inference is used in Throttler to avoid lots of type wildcards (e.g.,
? extends T). This allows static type checking, and avoids raw types and
unchecked casts. This refactoring makes the code more robust and more
readable.

Configurability is also improved. Currently, Throttler is hardcoded for loading
all resources. Though it is not a problem by itself, there are cases
where developers would like to use a different scheduling policy, or simply
load textures synchronously for development. In this patch, the API of the
Throttler is extracted as interface Scheduler, and the Scheduler is run-time
configurable. A sample scheduler SyncScheduler is provided, which loads
resources synchronously and can be used for debugging purposes.

Some static members and fields in the Throttler and Async* classes are
changed to non-static, and the classes are converted into Singletons. Therefore,
their initialization sequence can be better controlled. Now, they do not register
themselves until they are called by AsyncManager.