Don't add an object to the scene twice
GVRSceneObject.addChildObject will throw an exception if you try to add
a scene object that already has a parent.
Also made a common root object in GVRScene instead of the mChildren
array. Added GVRScene.getRoot() to get the hierarchy root. Perhaps we
can use the root constructively to attach the camera rig, etc.