Support for Google VR
This commit contains changes to support google vr:

https://github.com/googlevr/gvr-android-sdk.git

The google vr sdk is currently Android Studio only. This
change would not run with builds made on Eclipse.

The framework would now default to Google VR when Oculus is
not present on the device.

Monoscopic can still be used by forcing the mode using the
vr app settings.

The api versions have also been bumped from 19 to 21 to
support this change.