GVRFrustumPicker user-specified frustum, GVRPicker bug fixes
GVRPicker worked correctly when attached to the camera but not when
attached to a scene object. To fix this bug I changed the native picker
to use world coordinates instead of camera coordinates for the ray. This
allows any scene object to become the origin of the ray more easily.
GVRSphereCollider was not honoring the setRadius value so I fixed that.
I added support to GVRFrustumPicker for a user-defined frustum, not just
the current camera.