Fixed picking bug
The wrong matrix was being used for transforming the picking ray causing
it to pick the wrong thing.
Now the picker actually looks at the object it is attached to and gets
its position and orientation from that. Before it was incorrectly always
using the camera view matrix no matter what it was attached to. Some of
the demos have bugs because they attach the picker to the wrong object
and that used to work. I am doing a pull request for these demos.
Removed logging and exposed GVRCameraRigBase.getHeadTransformObject.