GVRAnimator and GVRExternalScene
Mihail requested components instead of GVRModelSceneObject. I agree so I
reduced dependency on GVRModelSceneObject (so we can take it out) and
introduced two new components.
GVRAnimator contains a list of animations (to replace the list that was
in GVRModelSceneObject) but it has more functionality. It can operate on
the animation list as a whole (set the repeat count, mode, start
animations)
GVRExternalScene is a component that contains a filename and is capable
of loading an asset at runtime. If you put this into the scene it
references an external asset that can be loaded programmatically, X3D
needs this to implement the anchor node.
This pull request is currently broken because getting the camera rig
crashes - do not merge yet!