Support for multitexturing and layered textures (#870)
* Support for multitexturing and layered textures

Updated the GVRPhongShader to support up to four sets of texture
coordinates and allow layering of two textures for each type (diffuse,
specular, ...)

Added GVRMaterial.setTexCoord to establish which set of texture
coordinates go with which texture

Updated Jassimp file importer to use multitexturing and layered texture
support

Changes to shaders:
a_tex_coord is no longer supported: use a_texcoord
vec4 a_texcoord wont work anymore either, use vec2 a_texcoord
All of the examples are updated in a separate pull request to make these
changes in the shaders

* Fixed problem with cubemap sample

The problem really wasn't about the cubemap shader. It was the post effect shader that GVRConsole uses to display the statistics that was broken. I fixed that. It was also getting a GL error trying to get texture coordinates that are not there. I put in a native Mesh accessor that will let you get texcoords without throwing an exception if they are not there.

* Fix conversion problem in Jassimp

* removed uniform block, enable batching, disable writing shaders to disk