Fix crash when loading vrbenchmark bmw.fbx
There was a check for texture coordinates in Mesh::generateVAO that will
crash if a mesh does not have texture coordinates. We used to supply
them always but now we honor the FBX file and this asset had a mesh
without texcoords.

This does NOT fix the synchronization issue in vrappbenchmark. This
application loads assets simultaneously from two threads. It now hangs
when loading textures. I traced this to an issue in the event manager.
GVREventManager.findHandlerMethod caches the method based on the event
name only. This means you can only have ONE HANDLER in GVRContext for a
particular event. Of course, GVRAssetLoader needs two if it is going to
run on two threads simultaneously...