Refactored shadowmap code, exposed render to texture (#1099)
* Refactored shadowmap code, exposed render to texture

This patch refactors the shadow mapping code from a hard-coded mess in
light.cpp into a set of general components that implements shadow
mapping. Now you can render the shadow maps because it is implemented on
top of RenderTexture.
1. Added new component GVRRenderTarget which will render to a texture
from the viewpoint of the scene object it is attached to. This component
renders to a GVRRenderTexture via a DrawFrameListener.
2. Added GVRContext.cullAndRender function to cull and render to a
GVRRenderTarget. This is what GVRRenderTarget calls each frame.
3. Added GVRShadowMap component which derives from GVRRenderTarget and
encapsulates the shadow mapping code which was formerly spewed all over
light.cpp
4. Added GVRRenderTextureArray to expose an array of render textures
(which is what shadow mapping uses).

* Add RenderTexture::setBackColor to set background color

RenderTexture now honors a background color and a flag indicating whether or not to use the stencil buffer. When rendering to a texture, RenderTarget will set the background color of the RenderTexture to match the camera.

* javadoc for GVRCamera.setViewMatrix, GVRDirectLight.setCastShadow ,GVRSpotLight.setCastShadow