Fix problem with picking test "canPickMeshWithObject" (#1129)
* Fix problem with picking test "canPickMeshWithObject"

This test creates a picker, starts listening for pick events and then attaches the picker to a scene object. This can cause false events to be generated between the time the picker starts listening and the time it is attached. This constructor eliminates the problem by attaching the picker first and then starting to listen for events.

* Fix picking from object

Default the pick ray  to (0,0,0,0,0,1) when picking from an object