Added START_ANIMATIONS import setting
The START_ANIMATIONS setting will cause animations from a loaded model to automatically start when the asset is added to the scene. (Currently X3D does not honor fhis flag).

Also made the "cacheEnabled" flag actually do something. It was not correctly implemented. Now if you clear this flag when loading the model it will NOT cache textures.