Try to automatically detect transparency (#1212)
* First take on transparency.

First take on automatically setting render order based on transparency
in the texture.

* Fixed logic when adjusting for transparency.

Fixed logic when adjusting for transparency.
Also added warning if we detect and RGBA image that is fully opaque.

* add check for diffuseTexture

add check for diffuseTexture

* fixes

fixes

* fix tga

fix tga for 24bit images (indicate that it doesn't have alpha)

* cleanup and set things up for transition to native

cleanup and set things up for transition to moving the hasTransparency()
    mechanism to the native side.

* off by one error in loop

had an off by one error in the loops.  fixed.