Fixed for embedded textures, texparams (#1327)
* Fixed for embedded textures, texparams


    made GVRAssetLoader.getDefaultTextureParameters() public
    fixed texture counting issue with embedded textures

* put back Roshan's fix

* Don't load a failed texture twice

When a texture fails to open, we make a pending request for it anyway which also fails to open. I eliminated the pending request for failed textures.
I also changed the code which makes the default 32x32 white texture to not load a resource but instead make the bitmap programatically.