fix for multiple texcoord sets in jassimp
GVRMaterial.setTexCoord would silently fail if you did not set a texture
beforehand. Really we should get rid of this API and just use
GVRTexture.setTexCoord directly - this does the real work. I changed it
to throw an exception if it doesn't have a texture so the user at least
knows about it. I also changed jassimp to call GVRTexture.setTexCoord.

This fix is highly localized to jassimp.