Eliminate GVRScene.bindShaders
It is no longer necessary to call bindShaders when objects are added to the scene. This function is called individually for each RenderData if the shader is not valid. It is better to defer binding the shader until it is known what type of RenderTarget it will be used with.