widgetplugin: rework so events propagate after pause/resume (#1669)
* AssetLoader: error checking

If a texture is loaded early during initialization and cannot be found, a default image is used. If the error occurs early enough, the default image isn't there yet.

GVRPhongLayeredShader is only capable of layering two textures. Now we check if the FBX file layers more than two and no longer crash the shader.
This patch checks for these cases.

* Cursor controller bug fixes

1. Fix cursor depth scaling in Gear controller
2. Set controller active state in Gaze controller
3. Combined GVRInputManager and GVRInputManagerImpl into one class
4. GVRPicker enables SEND_TO_HIT_OBJECT when using controller
5. Fix Gear controller duplicate motion events
Picking with the controller was wonky because the controller was
automatically generating motion events. GVRViewSceneObject and
GVRWidgetPlugin also convert controller events into motion events within
the quad representing the Android view. This patch suppresses the motion
events automatically generated by the controller which fixes picking
inside Android views. GVRGearController.sendEventsToActivity is added to
enable routing the events to the activity if this is desired.

* 3DCursor: handle cursor reassignment

The Gear controller comes on line later than the other controllers. This
patch makes the 3D cursor library recognize the controller and reassign
it to one of the active cursors. Previously an active cursor would never
get assigned to a different controller even if it was higher priority.

* Support controller in scene editor and widget plugin

Modified widget plugin to handle controller events. Clicking works but
scrolling inside the widget does not.

* 3dcursor: make some functions internal

These are only used by CursorManager.java

* constant cursor size, picking thread safety, keyboard cleanup

1. Modified GVRCursorController to keep the cursor a constant size as it
is projected on objects of various distances.
2. Lock around picking event generation
3. Fix GVRViewSceneObject event processing
4. Another try at fixing Widget plugin

* Fixed widget plugin event handling

finally picking seems to work with the widget plugin

* Added GVRCursorController.startDrag, stopDrag

Facilitates dragging with the cursor controller. These functions are synchronized around cursor movement so they are thread-safe.

* Added GVRCursorController.startDrag, stopDrag

Facilitates dragging with the cursor controller. These functions are synchronized around cursor movement so they are thread-safe.

* add extra node for cursor scaling

Extra node ensures a dragged object will not be scaled and that the initial scale on the cursor is preserved

* Cursor controller bug fixes

1. Fix cursor depth scaling in Gear controller
2. Set controller active state in Gaze controller
3. Combined GVRInputManager and GVRInputManagerImpl into one class
4. GVRPicker enables SEND_TO_HIT_OBJECT when using controller
5. Fix Gear controller duplicate motion events
Picking with the controller was wonky because the controller was
automatically generating motion events. GVRViewSceneObject and
GVRWidgetPlugin also convert controller events into motion events within
the quad representing the Android view. This patch suppresses the motion
events automatically generated by the controller which fixes picking
inside Android views. GVRGearController.sendEventsToActivity is added to
enable routing the events to the activity if this is desired.

* 3DCursor: handle cursor reassignment

The Gear controller comes on line later than the other controllers. This
patch makes the 3D cursor library recognize the controller and reassign
it to one of the active cursors. Previously an active cursor would never
get assigned to a different controller even if it was higher priority.

* Support controller in scene editor and widget plugin

Modified widget plugin to handle controller events. Clicking works but
scrolling inside the widget does not.

* 3dcursor: make some functions internal

These are only used by CursorManager.java

* constant cursor size, picking thread safety, keyboard cleanup

1. Modified GVRCursorController to keep the cursor a constant size as it
is projected on objects of various distances.
2. Lock around picking event generation
3. Fix GVRViewSceneObject event processing
4. Another try at fixing Widget plugin

* Fixed widget plugin event handling

finally picking seems to work with the widget plugin

* Added GVRCursorController.startDrag, stopDrag

Facilitates dragging with the cursor controller. These functions are synchronized around cursor movement so they are thread-safe.

* Added GVRCursorController.startDrag, stopDrag

Facilitates dragging with the cursor controller. These functions are synchronized around cursor movement so they are thread-safe.

* add extra node for cursor scaling

Extra node ensures a dragged object will not be scaled and that the initial scale on the cursor is preserved

* gvr-keyboardview: Fixed picking

1. use mesh collider instead of box collider
2. disable default picker for keyboard if picker is changed
3. move numeric keypad slightly close to viewer

* GVRCursorController.sendEventsToActivity

Turns out this function is needed with other controllers too

* widgetplugin: fix controller event processing

* add logging for event dispatch

* Fix swipe in 3D cursor settings view

* Add useControllerTypes to gvr.xml

useControllerTypes replaces useAndroidWearTouchpad and useGazeCursorController booleans with a list of controller types to enable. The list may contains the following words: gaze controller mouse gamepad geartouchpad external

* Allow GVRViewSceneObject to have a gesture detector

Adds a gesture detector to GVRViewSceneObject to more easily capture touch gestures on a view.

* Disable all controllers on startup

This fixes issues with apps that do not use controllers at all. They should be dormant in this case.

* 3dcursor: fix swiping in settings view

* scene-editor: useControllerTypes in gvr.xml

* Fix bug with removing and adding Gear controller

stop dragging when  cursor removed from scene

* framework: fix a NPE in GVRTouchPadGestureDetector; e1 and/or e2 can be null in onFling

not the root cause but just a workaround; root cause somewhere in the event generation

* widgetplugin: rework so events propagate after pause/resume

root cause must be some missing cleanup of gdx; but it is wasteful to
recreate all that stuff on pause/resume so avoiding all that; as a side
effect pause/resume works